requirements.txt
fastapi
uvicorn[standard]
scikit-learn
numpy
python-multipart
pytest
Pillow
joblib
pytest-cov